+++ cuckoo_index.h
@@ -29,9 +29,9 @@ namespace ci {
 
 class CuckooIndex : public IndexStructure {
  public:
-  bool StripeContains(size_t stripe_id, int value) const override;
+  bool StripeContains(size_t stripe_id, long value) const override;
 
-  Bitmap64 GetQualifyingStripes(int value, size_t num_stripes) const override;
+  Bitmap64 GetQualifyingStripes(long value, size_t num_stripes) const override;
 
   std::string name() const override { return name_; }
